requirements.txt
fastapi
uvicorn[standard]
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

transformers==4.44.2
google-generativeai
sentence-transformers==3.0.1

llama-index==0.11.19
llama-index-embeddings-huggingface==0.3.1
llama-index-vector-stores-faiss==0.2.1
llama-index-retrievers-bm25
faiss-cpu==1.8.0
google-generativeai
numpy<2.0
tqdm==4.66.4
datasets==2.20.0
pyyaml==6.0.2
playwright==1.47.0
streamlit==1.37.0
